machine_health: Reduce samplerate
Currently the conversion start pulse is generated at 1MHz. To be able to
read back the generated samples the SPI interface needs to run at 50MHz.
This is quite challenging to do over the PMOD interface simply because of
its electrical properties.

Reduce the samplerate to 50kHz since this is sufficient for this
application and it allows us to read the data much more reliable. Also
increase the width of the conversion start pulse from 10ns to 160ns for
similar reasons.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>