machine_health: Use AD7980 without busy indicator
The AD7980 has multiple different operating modes. Currently it is used in
the so called busy indicator mode. Unfortunately in this mode it requires
17 SCLK to release the SDO line and go back to tristate mode. This would
require a very specialized SPI controller that generates a extra SCLK cycle
after the 16 data bits have been read. So instead switch to non-busy
indicator mode.

For this mode the CNVST pulse needs to be wider than 710 ns, change it to
1280 ns to be on the safe side. In addition to that we need to generate the
trigger signal for the SPI transfer inside the FPGA rather than using a
signal supplied by the converter. For this the newly introduced
end_of_pulse signal of the util_pulse_gen block is used and the
util_sigma_delta block is removed from the design.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>