machine_health: Place SPI SCK and MISO registers in IOB
The SPI clock is generated in the FPGA and forwarded to the converter which
updates the MISO line when it sees a rising edge. This signal is forwarded
from the converter to the FPGA which captures it on the opposite edge.

If the total delay of propagation from the clock signal register inside the
FPGA to the converter and the MISO signal from the converter back to the
FPGA capture register is larger than half the SPI clock period the MISO
signal will arrive too late and the capture SPI word is shifted by one.

Placing the SCK output and MISO input register in the IOB makes sure that
the internal delay from the register to the IO pad is minimal. This allows
operation at a faster SPI clock frequency.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>